requirements.txt
fastapi
uvicorn
pydantic
python-jose
grpcio
grpcio-tools
pyyaml
docker
pytest
protobuf

# --- DB/ORM 및 환경변수 관리 ---
sqlalchemy
aiosqlite
python-dotenv
alembic
greenlet 

# --- 테스트 및 커버리지 ---
pytest-cov
pytest-asyncio
coverage

# --- 추가된 항목 ---
email-validator
dnspython
passlib[bcrypt]
httpx